requirements.txt
accelerate==0.19.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
einops==0.6.1
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
filelock==3.12.0
fonttools==4.39.3
freetype-py==2.4.0
future==0.18.3
huggingface-hub==0.0.8
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.5.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
moviepy==1.0.3
msgpack==1.0.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.6.0
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyglet==2.0.6
Pygments==2.15.1
pymeshlab==2022.2.post3
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
stack-data==0.6.2
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.10.3
tomlkit==0.11.0
# torch==1.10.1+cu111
# torchaudio==0.10.1+cu111
# torchsummary==1.5.1
# torchvision==0.11.2+cu111
tornado
tqdm==4.63.0
traitlets
transformers==4.7.0
trimesh==3.12.6
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.12.1
yacs==0.1.8
zipp==3.9.0